refactor: domain 패키지 내 클래스 패키지 구조 변경
