+++ src/main/java/racingcar/domain/car/MoveValueStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 public interface MoveValueStrategy {
     int createMoveValue();
