+++ src/main/java/racingcar/domain/car/Position.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import java.util.Objects;
