+++ src/test/java/racingcar/domain/attemptnumber/AttemptNumberTest.java
@@ -1,8 +1,9 @@
-package racingcar.domain;
+package racingcar.domain.attemptnumber;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.domain.attemptnumber.AttemptNumber;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
